Allow system and application folder locations to be configurable. Document usage in resources.
Have an environment that has split the system and application folders because of reasons (inherited legacy code). Couldn't use base configuration and had to write a "custom" kernel which copied `CodeIgniterKernel` and set those to what was needed and then needed to write a "custom" autoloader which was the `CodeIgniterClassLoader` file with the throw exception for not instance of `CodeIgniterKernel` removed.

Ability to set the folder paths gets rid of a lot of unnecessary code and is more maintainable.